% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uz5_recap_annual_ranef.R
\name{compare_plot_annual_ranef}
\alias{compare_plot_annual_ranef}
\title{Compare plot that handles random effect models}
\usage{
compare_plot_annual_ranef(..., names = NULL)
}
\arguments{
\item{...}{two or a moult or moultmcmc model}

\item{names}{optional character vector of model names}
}
\value{
a plot comparing parameter estimates and their uncertainties
}
\description{
Compare plot that handles random effect models
}
