% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{siskin_capture}
\alias{siskin_capture}
\title{Empirical moult dependend capture probabilities}
\format{
A data frame with 1000 rows and 2 variables:
\describe{
\item{pfmg}{pfmg}
\item{p_cap}{capture probability}
}
}
\source{
P. Boersch-Supan / Hugh Insley
}
\usage{
siskin_capture
}
\description{
A lookup dataset containing empirical capture probabilities for siskins
}
\keyword{datasets}
