% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moultmcmcExtra-package.R
\docType{package}
\name{moultmcmcExtra-package}
\alias{moultmcmcExtra}
\alias{moultmcmcExtra-package}
\title{The 'moultmcmcExtra' package.}
\description{
Add-on functionality for moultmcmc
}
\author{
\strong{Maintainer}: Philipp Boersch-Supan \email{pboesu@gmail.com} (\href{https://orcid.org/0000-0001-6723-6833}{ORCID})

Authors:
\itemize{
  \item Philipp Boersch-Supan \email{pboesu@gmail.com} (\href{https://orcid.org/0000-0001-6723-6833}{ORCID})
}

}
\keyword{internal}
