% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moultmcmcExtra-package.R
\docType{package}
\name{moultmcmcExtra-package}
\alias{moultmcmcExtra-package}
\alias{moultmcmcExtra}
\title{The 'moultmcmcExtra' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.26.11. https://mc-stan.org
}
